requirements.txt
pandas>=1.5.0
numpy>=1.23.0
ortools>=9.0
pyyaml>=6.0
pyarrow>=10.0.0
openpyxl>=3.0.0

